Move pipeline CPU models & components to dinocpu.pipelined package
